+++ src/ensureError.ts
@@ -1,3 +1,8 @@
+/**
+ * Ensures given value is an instance of Error.
+ * @param {*} e - value to check
+ * @returns Error - original error or new Error instance
+ */
 const ensureError = (e: unknown) => {
     if (e instanceof Error) {
         return e;