+++ src/isEmpty.ts
@@ -8,15 +8,16 @@ interface ObjectWithSize {
 
 /**
  * Returns true if passed argument seems to be empty.
- * Primitives (excluding string) are always empty (even truthy).
- * Only empty strings are considered empty.
+ * Nil values are empty.
+ * Strings are considered empty when length is 0.
+ * Other primitives will throw an error.
  * Objects are considered empty when doesn't have any enumerable & own property.
  * Arrays and array-like objects are considered empty when length value is 0.
  * Map, Set and -like objects are considered empty when size value is 0.
  *
  * @param {*} obj - source value
  * @example isEmpty({}) // true
- * @example isEmpty(100) // true
+ * @example isEmpty(100) // throws
  * @example isEmpty([]) // true
  * @example isEmpty([1]) // false
  * @example isEmpty({ length: 5 }) // false
@@ -28,8 +29,11 @@ const isEmpty = (obj: unknown) => {
     if (typeof obj === "string") {
         return !obj.length;
     }
-    if (typeof obj !== "object" || obj === null) {
+    if (obj == null) {
         return true;
+    }
+    if (typeof obj !== "object") {
+        throw new Error("isEmpty cannot be used on primitives");
     }
     if (Array.isArray(obj)) {
         return !Object.keys(obj).length;