+++ src/merge.ts
@@ -15,7 +15,7 @@ interface Merge {
 
 /**
  * Shallow merges given objects into new object. It does not mutate any object given. Allows removing properties as
- * well by assigning special merge.UNSET value.
+ * well by assigning special `merge.UNSET` value.
  * @param {...object} args - input objects
  * @returns {*}
  */