+++ src/mostFrequent.spec.ts
@@ -18,7 +18,7 @@ describe("mostFrequent", () => {
     });
 
     it("compares strictly", () => {
-        mostFrequent([1, 2, "1", 1, 2, 2]).must.equal(2);
+        mostFrequent([1, "1", 1, "1", 2, 2, 2]).must.equal(2);
     });
 
     it("works with NaNs", () => {