+++ src/waitSync.ts
@@ -1,0 +1,14 @@
+/**
+ * Synchronously wait for a given time, blocking the event loop [!]
+ * @param {number} timeMs - time to wait
+ * @returns {Promise<void>}
+ */
+const waitSync = (timeMs = 0) => {
+    const s = Date.now();
+    // eslint-disable-next-line no-empty
+    while (Date.now() - s < timeMs) {}
+};
+
+export {
+    waitSync,
+};